Update-de11f831-a6d9-4595-acb8-6eb7e9ce0695

+++ src/main/java/com/bss/inc/cameraon/utils/net/angryscan/feeders/Feeder.java
@@ -19,7 +19,7 @@ public interface Feeder extends Plugin {
     /**
      * @return value from 0 to 100, describing the amount of work already done
      */
-    double percentageComplete();
+    int percentageComplete();
 
     /**
      * @return information about feeder's current settings.
